backend/pyproject.toml
[project]
name = "verity-backend"
version = "0.1.0"
description = "Evidence-based health claim verification system using LangGraph and Claude"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Core framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    # LangGraph & LangChain
    "langgraph>=0.2.62",
    "langchain>=0.3.14",
    "langchain-anthropic>=0.3.3",
    "langgraph-checkpoint-sqlite>=2.0.8",
    # Database
    "sqlalchemy>=2.0.36",
    "aiosqlite>=0.20.0",
    "alembic>=1.14.0",
    # PubMed integration
    "biopython>=1.84",
    "httpx>=0.28.1",
    # Observability
    "langsmith>=0.2.7",
    # Utilities
    "pydantic>=2.10.6",
    "pydantic-settings>=2.7.1",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.20",
    "langchain-groq>=1.1.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.2",
    "httpx>=0.28.1",
    "ruff>=0.8.6",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
